refactor(SingleNumber): InputNumberFactory에 몰려있던 검증 로직을 SingleNumber로 분리
InputNumberFactory에서 음수를 판단하는 등, 숫자 값 객체가 해야할 일을 한다고 생각이 들었습니다. 그래서 일부를 SingleNumber 객체로 옮겨 적용하였습니다.
