+++ src/main/java/baseball/domain/number/inputnumber/InputNumberFactory.java
@@ -9,7 +9,8 @@ import java.util.stream.IntStream;
 public class InputNumberFactory {
 
     public static InputNumbers newInstance(String inputValue) {
-        validateInputValue(inputValue);
+        validateLengthForGame(inputValue);
+        validateDuplicate(inputValue);
         return new InputNumbers(generateSingleNumberList(inputValue));
     }
 
@@ -26,31 +27,11 @@ public class InputNumberFactory {
                 .collect(Collectors.toList());
     }
 
-    // TODO SingleNumber에서 구현
-    private static void validateInputValue(String inputValue) {
-        validateLengthForGame(inputValue);
-        validatePositiveNumber(inputValue);
-        validateContainsZero(inputValue);
-        validateDuplicate(inputValue);
-    }
-
     private static void validateLengthForGame(String inputValue) {
         int length = inputValue.length();
 
         if (length != 3) {
             throw new IllegalArgumentException("세자리의 숫자를 입력해 주세요.");
-        }
-    }
-
-    private static void validatePositiveNumber(String inputValue) {
-        if (inputValue.contains("-")) {
-            throw new IllegalArgumentException("양수만 입력해 주세요.");
-        }
-    }
-
-    private static void validateContainsZero(String inputValue) {
-        if (inputValue.contains("0")) {
-            throw new IllegalArgumentException("0은 포함될 수 없습니다. 다시 입력해 주세요.");
         }
     }
